Fix previous commit
* split .h files into .h and .cpp files
* added files to CMakeLists.txt
* fixed typo in controller.h for moveRight
